+++ zoo_keeper/recipes/__init__.py
@@ -15,6 +15,18 @@ def get(species):
         from . import simple_car as mod
     elif species == "filing_cabinet":
         from . import filing_cabinet as mod
+    elif species == "table":
+        from . import table as mod
+    elif species == "crt_tv":
+        from . import crt_tv as mod
+    elif species == "atm":
+        from . import atm as mod
+    elif species == "vending_machine":
+        from . import vending_machine as mod
+    elif species == "briefcase":
+        from . import briefcase as mod
+    elif species == "cash_stack":
+        from . import cash_stack as mod
     else:
         raise KeyError(f"No recipe for species '{species}'")
     return mod.build